pyproject.toml
[tool.poetry]
name = "expense-sorter"
version = "0.1.0"
description = ""
authors = ["Quinten Rosseel <quinten.rosseel@hotmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
default = true

[tool.poetry.dependencies]
torch = "2.0.0"  # 2.4 doesn't work. 
python = "^3.10"
pymupdf = "^1.24.9"
pillow = "^10.4.0"
pillow-heif = "^0.18.0"
pytest = "^8.3.2"
transformers = "^4.44.0"
timm = "^1.0.8"  # For the DetrFeatureExtractor
numpy = "1.26.4"  # For the DetrFeatureExtractor
matplotlib = "^3.9.1.post1"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
